Support plugins in the fetch cell
