+++ src/editor/actions/sagas/language-plugin-saga.js
@@ -1,0 +1,44 @@
+import CodeMirror from "codemirror";
+import { call, put } from "redux-saga/effects";
+
+import {
+  addLoadingLanguageMsgToHistory,
+  addPluginParseErrorToHistory
+} from "../console-message-actions";
+import { triggerEvalFrameTask } from "./eval-frame-sender";
+
+export const languageReady = (state, lang) =>
+  Object.keys(state.loadedLanguages).includes(lang);
+
+export const languageKnown = (state, lang) =>
+  Object.keys(state.languageDefinitions).includes(lang);
+
+export const languageNeedsLoading = (state, lang) =>
+  languageKnown(state, lang) && !languageReady(state, lang);
+
+export function* loadLanguagePlugin(pluginData) {
+  yield call(triggerEvalFrameTask, "EVAL_LANGUAGE_PLUGIN", {
+    pluginData
+  });
+  // FIXME this empty function argument seems sketchy
+  yield call([CodeMirror, "requireMode"], pluginData.codeMirrorMode, () => {});
+  yield put({
+    type: "ADD_LANGUAGE_TO_EDITOR",
+    languageDefinition: pluginData
+  });
+}
+
+export function* loadKnownLanguage(pluginData) {
+  yield put(addLoadingLanguageMsgToHistory(pluginData.displayName));
+  yield call(loadLanguagePlugin, pluginData);
+}
+
+export function* evaluateLanguagePlugin(pluginText) {
+  try {
+    const pluginData = JSON.parse(pluginText);
+    yield call(loadLanguagePlugin, pluginData);
+  } catch (error) {
+    yield put(addPluginParseErrorToHistory(error.message));
+    throw error;
+  }
+}
